// sisref-v2-dashboard-client/src/app/elementos/interfaces/IRefeicaoDoDoHistorico.ts
export interface IRefeicaoDoHistorico {
    turno: 1 | 2 | 3 | 4;
    cardapio: {
        id: number;
        date: string;
        description: string;
        campus_id: number;
        meal_id: number;
    };
    refeicao: {
        id: number;
        description: string;
        timeEnd: string;
        timeStart: string;
        campus_id: number;
        qtdTimeReservationEnd: number;
        qtdTimeReservationStart: number;
    };
    status: "a-ser-utilizado" | "utilizado" | "cancelado" | "nao-utilizado";
}